Generic Code; Tilemap
Created a more general script for levels so I can build off of that for every level. Also changed the potion picking-up and resetting when the player dies so that only the potions picked up during the level get removed, added more details to the tilemap, and made the map bigger.