Started Adding Enemies; Fixing a Bug
Started adding enemies to the game. Also made the spell scene more efficient by creating a CollisionShape2D within the scene itself instead of having to add it to every instance of the scene in each level. While doing this, however, I managed to create a bug where the spells' CollisionShape2D is always the one for the fireball, even if the spell is not a fireball, so now I am trying to fix this.