Updated Tilemap/Map; Fixed Errors
Added to the tilemap/map for each level, and made numerous small fixes. For example, I fixed the direction the spells' sprites are facing, and several issues related to the potions, including the fact that the number of potions goes to 0 when the player dies when I only want it to decrease by the number of potions the player picked up during that scene/level.